processing benchmarking PPI data

+++ BERT-Relation-Extraction/main_task.py
@@ -15,6 +15,7 @@ import logging
 from argparse import ArgumentParser
 import os
 import numpy as np
+import csv
 
 '''
 This fine-tunes the BERT model on SemEval, FewRel tasks
@@ -26,7 +27,10 @@ logger = logging.getLogger('__file__')
 
 if __name__ == "__main__":
 	parser = ArgumentParser()
-	parser.add_argument("--task", type=str, default='semeval', help='semeval, biocreative, fewrel') # biocreative added. 11-26-2020
+	# [GP][START] - BioCreative added. 11-26-2020
+	#			  - AImed, BioInfer, HPRD50, IEPA, LLL added. 02-19-2021
+	parser.add_argument("--task", type=str, default='semeval', help='semeval, fewrel, BioCreative, BioCreative_BNL, AImed, BioInfer, HPRD50, IEPA, LLL, AImed_BioInfer_HPRD50_IEPA_LLL')
+	# [GP][END] - BioCreative added. 11-26-2020
 	parser.add_argument("--train_data", type=str, default='./data/SemEval2010_task8_all_data/SemEval2010_task8_training/TRAIN_FILE.TXT', \
 						help="training data .txt file path")
 	parser.add_argument("--test_data", type=str, default='./data/SemEval2010_task8_all_data/SemEval2010_task8_testing_keys/TEST_FILE_FULL.TXT', \
@@ -51,17 +55,22 @@ if __name__ == "__main__":
 	parser.add_argument("--infer", type=int, default=1, help="0: Don't infer, 1: Infer")
 	
 	# [GP][START] - arguments for cross-validation. 11-29-2020
-	parser.add_argument("--num_samples", type=int, default=300, help="No of samples for evaluation")
+	parser.add_argument("--num_samples", type=int, default=-1, help="No of samples for evaluation. -1 means to use all samples.")
 	parser.add_argument("--do_cross_validation", action="store_true", help="Whether to use cross-validation for evaluation.")
 	parser.add_argument("--num_of_folds", default=10, type=int, help="The number of folds for the cross validation.")
+	parser.add_argument("--ratio", type=str, default='60-20-20', help="train/dev/test ratio: 80-10-10, 70-15-15, 60-20-20")
 	parser.add_argument("--eval_value", type=str, help="best: get the best value among epochs, last: get the value of last epoch")
 	parser.add_argument("--do_one_class_classification", action="store_true", help="Whether to use sigmoid for outputs rather than softmax. \
 																					this is used for testing BioCreative 'PPIm' label.")
 	parser.add_argument("--threshold", default=0.5, type=float, help="threshold value to determine one-class outputs. \
 																	  this is used together with do_one_class_classification.")																	
-	parser.add_argument("--use_predefined_cls", action="store_true", help="Whether to use predefined labels such as PPIm.")
+	parser.add_argument("--predefined_cls", type=int, help="predefined labels such as PPIm.")
 	parser.add_argument("--result_dir", type=str, help="result directory path")
 	# [GP][END] - arguments for cross-validation. 11-29-2020
+	
+	# [GP][START] - added relation mapper path for infer task. 02-27-2021
+	parser.add_argument("--rm_file", type=str, help="(for infer) relation mapper path")	
+	# [GP][END] - added relation mapper path for infer task. 02-27-2021
 	
 	args = parser.parse_args()
 	
@@ -71,9 +80,13 @@ if __name__ == "__main__":
 		if args.do_cross_validation:
 			num_of_datasets = args.num_of_folds
 		
-		if not os.path.exists(args.result_dir + '/' + str(args.num_samples)):
-			os.makedirs(args.result_dir + '/' + str(args.num_samples))
-		
+		if args.num_samples == -1:
+			if not os.path.exists(args.result_dir + '/all'):
+				os.makedirs(args.result_dir + '/all')
+		else:
+			if not os.path.exists(args.result_dir + '/' + str(args.num_samples)):
+				os.makedirs(args.result_dir + '/' + str(args.num_samples))
+			
 		if args.model_no == 0:
 			result_file = 'BERT_'
 		elif args.model_no == 1:
@@ -85,8 +98,11 @@ if __name__ == "__main__":
 			result_file += 'with_MTB_cv_result.txt'
 		else:
 			result_file += 'without_MTB_cv_result.txt'
-			
-		result_file = os.path.join(args.result_dir + '/' + str(args.num_samples), result_file)
+		
+		if args.num_samples == -1:
+			result_file = os.path.join(args.result_dir + '/all', result_file)
+		else:
+			result_file = os.path.join(args.result_dir + '/' + str(args.num_samples), result_file)
 		
 		if args.eval_value == 'best':
 			best_f1_per_cv_set = []
@@ -164,12 +180,13 @@ if __name__ == "__main__":
 				last_f1_accuracy_per_cv_set.append(test_result_per_epoch[-1]['accuracy'])
 			"""	
 			
-			# remove files for the next CV set.
-			os.remove(os.path.join("./data/" , "task_test_model_best_%d.pth.tar" % args.model_no))
-			os.remove(os.path.join("./data/" , "task_test_checkpoint_%d.pth.tar" % args.model_no))
-			os.remove(os.path.join("./data/" , "task_test_losses_per_epoch_%d.pkl" % args.model_no))
-			os.remove(os.path.join("./data/" , "task_train_accuracy_per_epoch_%d.pkl" % args.model_no))
-			os.remove(os.path.join("./data/" , "task_test_f1_per_epoch_%d.pkl" % args.model_no))
+			if args.do_cross_validation:
+				# remove files for the next CV set.
+				os.remove(os.path.join("./data/" , "task_test_model_best_%d.pth.tar" % args.model_no))
+				os.remove(os.path.join("./data/" , "task_test_checkpoint_%d.pth.tar" % args.model_no))
+				os.remove(os.path.join("./data/" , "task_test_losses_per_epoch_%d.pkl" % args.model_no))
+				os.remove(os.path.join("./data/" , "task_train_accuracy_per_epoch_%d.pkl" % args.model_no))
+				os.remove(os.path.join("./data/" , "task_test_f1_per_epoch_%d.pkl" % args.model_no))
 
 		if args.eval_value == 'best':
 			best_f1_accuracy_per_cv_set = np.array(best_f1_accuracy_per_cv_set)
@@ -211,17 +228,62 @@ if __name__ == "__main__":
 		
 	if (args.infer == 1) and (args.task != 'fewrel'):
 		inferer = infer_from_trained(args, detect_entities=True)
-		test = "The surprise [E1]visit[/E1] caused a [E2]frenzy[/E2] on the already chaotic trading floor."
+		#test = "The surprise [E1]visit[/E1] caused a [E2]frenzy[/E2] on the already chaotic trading floor."
+		test = "[E1]PCLI[/E1] was also significantly correlated to [E2]FOXM1[/E2] and CKS1B expression , but at a lower extent ."
 		inferer.infer_sentence(test, detect_entities=False)
-		test2 = "After eating the chicken, he developed a sore throat the next morning."
-		inferer.infer_sentence(test2, detect_entities=True)
+		test = "[E1]PCLI[/E1] was also significantly correlated to FOXM1 and [E2]CKS1B[/E2] expression , but at a lower extent ."
+		inferer.infer_sentence(test, detect_entities=False)
+		#test2 = "After eating the chicken, he developed a sore throat the next morning."
+		#inferer.infer_sentence(test2, detect_entities=True)
 		
+		pos_pairs = {}
+		neg_pairs = {}
+		with open(args.test_data) as fp:
+			lines = fp.readlines()
+			for sent in lines:
+			
+				pair = [x.upper() for x in sent.split() if '[/E' in x]
+				pair = [x.replace('[E1]', '').replace('[/E1]', '').replace('[E2]', '').replace('[/E2]', '') for x in pair]
+				pair.sort()
+				pair = '-'.join(pair)
+			
+				pred = inferer.infer_sentence(sent, detect_entities=False)
+				
+				#print(pair)
+				#print(pred)
+				
+				if pred == 0:
+					if pair in pos_pairs:
+						pos_pairs[pair] += 1
+					else:
+						pos_pairs[pair] = 1
+				elif pred == 1:
+					if pair in neg_pairs:
+						neg_pairs[pair] += 1
+					else:
+						neg_pairs[pair] = 1
+		
+		with open(os.path.join(args.result_dir, 'pos_ppi_count.csv'), mode='w') as f:
+			fw = csv.writer(f)
+			fw.writerow(['PPI', 'Count'])
+
+			for ppi, cnt in pos_pairs.items():
+				fw.writerow([ppi, cnt])
+		
+		with open(os.path.join(args.result_dir, 'neg_ppi_count.csv'), mode='w') as f:
+			fw = csv.writer(f)
+			fw.writerow(['PPI', 'Count'])
+
+			for ppi, cnt in neg_pairs.items():
+				fw.writerow([ppi, cnt])
+		
+		'''
 		while True:
 			sent = input("Type input sentence ('quit' or 'exit' to terminate):\n")
 			if sent.lower() in ['quit', 'exit']:
 				break
 			inferer.infer_sentence(sent, detect_entities=False)
-	
+		'''
 	if args.task == 'fewrel':
 		fewrel = FewRel(args)
 		meta_input, e1_e2_start, meta_labels, outputs = fewrel.evaluate()
